refactor: flatten editor package folder structure
Remove the ce/ee split in @plane/editor and consolidate everything under
src/core, matching the layout used in the enterprise repository.

- Move ce components, constants, extensions, helpers, and types into core
- Rename ce/extensions/slash-commands.tsx to additional-slash-command-options.tsx
  to avoid clashing with the existing slash-commands/ directory
- Merge ce type stubs and parser helpers into their core counterparts
- Drop the ee re-export shim (no consumers)
- Rewrite @/plane-editor/* imports to @/* and trim tsconfig path aliases
