--- packages/editor/src/ce/helpers/parser.ts
@@ -1,25 +1,0 @@
-/**
- * Copyright (c) 2023-present Plane Software, Inc. and contributors
- * SPDX-License-Identifier: AGPL-3.0-only
- * See the LICENSE file for details.
- */
-
-/**
- * @description function to extract all additional assets from HTML content
- * @param htmlContent
- * @returns {string[]} array of additional asset sources
- */
-export const extractAdditionalAssetsFromHTMLContent = (_htmlContent: string): string[] => [];
-
-/**
- * @description function to replace additional assets in HTML content with new IDs
- * @param props
- * @returns {string} HTML content with replaced additional assets
- */
-export const replaceAdditionalAssetsInHTMLContent = (props: {
-  htmlContent: string;
-  assetMap: Record<string, string>;
-}): string => {
-  const { htmlContent } = props;
-  return htmlContent;
-};
