--- packages/editor/src/ce/types/index.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) 2023-present Plane Software, Inc. and contributors
- * SPDX-License-Identifier: AGPL-3.0-only
- * See the LICENSE file for details.
- */
-
-export * from "./issue-embed";
-export * from "./editor-extended";
-export * from "./config";
