+++ packages/editor/src/core/components/document-editor-side-effects.ts
