+++ packages/editor/src/core/components/editors/document/editor.tsx
@@ -20,7 +20,7 @@ import { getEditorClassNames } from "@/helpers/common";
 // hooks
 import { useEditor } from "@/hooks/use-editor";
 // plane editor extensions
-import { DocumentEditorAdditionalExtensions } from "@/plane-editor/extensions";
+import { DocumentEditorAdditionalExtensions } from "@/extensions/document-extensions";
 // types
 import type { EditorRefApi, IDocumentEditorProps } from "@/types";
