+++ packages/editor/src/core/constants/assets.ts
