+++ packages/editor/src/core/constants/extensions.ts
