+++ packages/editor/src/core/extensions/additional-slash-command-options.tsx
