+++ packages/editor/src/core/extensions/core/extensions.ts
