+++ packages/editor/src/core/extensions/core/without-props.ts
