+++ packages/editor/src/core/extensions/unique-id/extension.ts
@@ -11,7 +11,7 @@ import type { Transaction } from "@tiptap/pm/state";
 import { v4 as uuidv4 } from "uuid";
 // constants
 import { CORE_EXTENSIONS, BLOCK_NODE_TYPES } from "@/constants/extension";
-import { ADDITIONAL_BLOCK_NODE_TYPES } from "@/plane-editor/constants/extensions";
+import { ADDITIONAL_BLOCK_NODE_TYPES } from "@/constants/extensions";
 import { createUniqueIDPlugin } from "./plugin";
 import { createIdsForView } from "./utils";
 // plane imports
