+++ packages/editor/src/core/helpers/asset-duplication.ts
