+++ packages/editor/src/core/hooks/use-collaborative-editor.ts
@@ -15,7 +15,7 @@ import { HeadingListExtension, SideMenuExtension } from "@/extensions";
 // hooks
 import { useEditor } from "@/hooks/use-editor";
 // plane editor extensions
-import { DocumentEditorAdditionalExtensions } from "@/plane-editor/extensions";
+import { DocumentEditorAdditionalExtensions } from "@/extensions/document-extensions";
 // types
 import type {
   TCollaborativeEditorHookProps,
