+++ packages/editor/src/core/types/editor-extended.ts
