+++ packages/editor/src/core/types/index.ts
@@ -15,5 +15,5 @@ export * from "./hook";
 export * from "./mention";
 export * from "./slash-commands-suggestion";
 export * from "./document-collaborative-events";
-
-export * from "@/plane-editor/types";
+export * from "./issue-embed";
+export * from "./editor-extended";
