+++ packages/editor/src/core/types/storage.ts
