--- packages/editor/src/ee/types/index.ts
@@ -1,7 +1,0 @@
-/**
- * Copyright (c) 2023-present Plane Software, Inc. and contributors
- * SPDX-License-Identifier: AGPL-3.0-only
- * See the LICENSE file for details.
- */
-
-export * from "src/ce/types";
